Save client id and secret when using grive -a
If you wanted to give a different than the default client id and
secret, you always had to give that on the commandline. Now it's saved
in the .give file. If given on the commandline, it's now also only
used when 'auth/-a' is given, in normal runs it's not anymore valid (and errors out).